backend/requirements.txt
Django==5.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
# psycopg2-binary==2.9.9 - Install this manually if needed
# Pillow==10.1.0 - Install this manually if neededpyotp==2.9.0
qrcode==8.2
